Check the findPoint def
